Remove ActionButton, HandleClickDown/Up, don't pass coords to KeyDown/Up
(should always be obtained from input device)
Gtk2+Mac, use 101=right mouse button; Win32 TODO; iPhone ???
Also change int iTime param to KeyDown/KeyUp into an unsigned long.

Rm corresponding "ExecuteCommand" method, and associated Gtk2 methods
 - preserving some that are needed for standard commands like "Save"...

+++ Src/DasherCore/ButtonMultiPress.cpp
@@ -27,7 +27,7 @@ CButtonMultiPress::CButtonMultiPress(CSettingsUser *pCreator, CDasherInterfaceBa
   : CDynamicButtons(pCreator, pInterface, pFramerate, iID, szName) {
 }
 
-void CButtonMultiPress::KeyDown(int iTime, int iId, CDasherView *pView, CDasherInput *pInput, CDasherModel *pModel, CUserLogBase *pUserLog) {
+void CButtonMultiPress::KeyDown(unsigned long iTime, int iId, CDasherView *pView, CDasherInput *pInput, CDasherModel *pModel, CUserLogBase *pUserLog) {
 
   if (m_bKeyDown) return;
